% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPReg.R
\name{BPReg}
\alias{BPReg}
\title{Bayesian penalized regressions for non-stationary Gaussian linear mixed effects model.}
\usage{
BPReg(
  y,
  x,
  time,
  id,
  n.samples,
  r.init,
  delta.init,
  method = c("BEN", "BAL", "BR", "BL")
)
}
\arguments{
\item{y}{A vector for the response variable.}

\item{x}{A data frame that includes scaled covariates whose coefficients will be penalized.}

\item{time}{A vector for the time variable (e.g. age)}

\item{id}{A vector for the individuals id.}

\item{n.samples}{MCMC sample size.}

\item{r.init}{value assumed for r, a paramater of the Gamma prior (see Gecili et al.2020).}

\item{delta.init}{value assumed for delta, a paramater of the Gamma prior (see Gecili et al.2020)}

\item{method}{It allows for BEN (Bayesian elastic-net), BAL (Bayesian adaptive lasso), BL (Bayesian lasso), and BR (Bayesian ridge).}
}
\value{
Returns posterior estimates and credible intervals.
}
\description{
Bayesian penalized regressions for non-stationary Gaussian linear mixed effects model.
}
\examples{
BPReg(y,x,time,id,n.samples =25000,r.init = 1,delta.init = 1,method = "BR")
}
