% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_aburn.R
\name{MCMC_aburn}
\alias{MCMC_aburn}
\title{Performs burnin to MCMC output object}
\usage{
MCMC_aburn(burnin)
}
\arguments{
\item{burnin}{amount of burnin.}
}
\value{
Returns remaining MCMC samples after burnin.
}
\description{
Performs burnin and returns remaining MCMC samples example.
}
\examples{
M=MCMC_aburn(burnin=5000)
}
