% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inf_criteria.R
\name{MCMCsummary}
\alias{MCMCsummary}
\title{MCMC summary}
\usage{
MCMCsummary(M)
}
\arguments{
\item{M}{Final MCMC output after burnin.}
}
\value{
Returns summary of posterior distributions of all parameters.
}
\description{
Returns summary of posterior distributions of all parameters.
}
\examples{
MCMCsummary(M)
}
