% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inf_criteria.R
\name{posterior_summary}
\alias{posterior_summary}
\title{Returns posterior summary}
\usage{
posterior_summary(M)
}
\arguments{
\item{M}{Final MCMC output after burnin.}
}
\value{
Returns posterior summary.
}
\description{
Summarizes posterior samples based on point estimates (mean or median), estimation errors (SD or MAD) and quantiles If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead.
}
\examples{
posterior_summary(M, probs = c(0.025, 0.975),robust =FALSE)
}
