% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inf_criteria.R
\name{Inf_criteria}
\alias{Inf_criteria}
\title{Inf_criteria}
\usage{
Inf_criteria(M)
}
\arguments{
\item{M}{Final MCMC output after burnin.}
}
\value{
AIC, BIC, log-likelihood.
}
\description{
Returns AIC, BIC, log-likelihood.################### posterior summarization ################################################################## posterior_summary: Summarizes posterior samples based on point estimates (mean or median), estimation errors (SD or MAD) and quantiles If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead. example: posterior_summary(M, probs = c(0.025, 0.975),robust =FALSE). MCMCplot: returns caterpillar plot for selected parameters with credible intervals. Points represent posterior medians. Thick lines represent 50 percent credible intervals while thin lines represent 95 percent credible intervals. ref_ovl = TRUE make circle empty #rank=TRUE orders. example: MCMCplot(M[,c(2:dim(x)\link{2})],params=colnames(M)[2:dim(x)\link{2}],ref_ovl = TRUE). MCMCtrace: returns a panel of trace and density plots for selected parameters. example: MCMCtrace(M, params = 'Int',ISB = FALSE,pdf = FALSE). MCMCsummary: returns summary of posterior distributions of all parameters. example: MCMCsummary(M)
}
\examples{
Inf_criteria(M)
}
