% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_aburn.R
\name{MCMC_aburn}
\alias{MCMC_aburn}
\title{Bayesian penalized regressions for non-stationary Gaussian linear mixed effects model.}
\usage{
MCMC_aburn(burnin)
}
\arguments{
\item{burnin}{amount of burnin.}
}
\value{
Performs burnin and returns remaining MCMC samples example.
}
\description{
Bayesian penalized regressions for non-stationary Gaussian linear mixed effects model.
}
\examples{
M=MCMC_aburn(burnin=40)
}
